refactored to benchmark/common
