+++ host_core/test/test_helper.exs
@@ -1,3 +1,4 @@
+Code.require_file("benchmark/common.exs", __DIR__)
 Code.require_file("support/event_watcher.exs", __DIR__)
 Code.require_file("support/common.exs", __DIR__)
 Code.require_file("support/constants.exs", __DIR__)
